+++ OneBrainSoul/Assets/Scripts/AI/Combat/Contexts/TeleportMobilityEnemyContext.cs
@@ -8,7 +8,7 @@ namespace AI.Combat.Contexts
         private bool _hasReachedDestination;
         
         protected TeleportMobilityEnemyContext(EntityType entityType, uint totalHealth, uint maximumHeadYawRotation, 
-            float radius, float height, float sightMaximumDistance, float fov, Transform headAgentTransform, Transform bodyAgentTransform) : 
+            float radius, float height, float sightMaximumDistance, uint fov, Transform headAgentTransform, Transform bodyAgentTransform) : 
             base(entityType, totalHealth, maximumHeadYawRotation,radius, height, sightMaximumDistance, 
             fov, headAgentTransform, bodyAgentTransform)
         {
